src/requirements.txt
python-telegram-bot==21.6
Flask==3.0.3
requests==2.32.3
httpx==0.27.0
openai==2.7.2
python-dotenv==1.1.1